requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==36.0.0
Django==3.2.10
django-cors-headers==3.10.1
django-dotenv==1.4.2
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mssql-django==1.1
openapi-codec==1.3.2
packaging==21.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyodbc==4.0.32
pyparsing==2.4.7
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
